requirements.txt
PyQt6>=6.4.0
selenium>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=1.3.0
python-dotenv>=1.0.0
webdriver-manager>=3.8.0
lxml>=4.9.3 